0812 Q8~Q14

+++ one star/10. Summing Digits.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 
 int main() {
-    long long int n ;
+    int n ;
     while (cin >> n) {
         if (n == 0) {
             return 0 ;
